explorer,views: fix charts page target ticket pool size
In (*explorerUI).Charts get target pool size from
exp.pageData.HomeInfo.PoolInfo.Target, and pass to page template.

explorer/types: change TicketPoolInfo.Target uint16=>uint32

+++ explorer/explorer.go
@@ -340,7 +340,7 @@ func New(cfg *ExplorerConfig) *explorerUI {
 				MeanVotingBlocks: exp.MeanVotingBlocks,
 			},
 			PoolInfo: types.TicketPoolInfo{
-				Target: exp.ChainParams.TicketPoolSize * exp.ChainParams.TicketsPerBlock,
+				Target: uint32(exp.ChainParams.TicketPoolSize) * uint32(exp.ChainParams.TicketsPerBlock),
 			},
 		},
 	}
@@ -492,7 +492,7 @@ func (exp *explorerUI) Store(blockData *blockdata.BlockData, msgBlock *wire.MsgB
 	// If BlockData contains non-nil PoolInfo, copy values.
 	p.HomeInfo.PoolInfo = types.TicketPoolInfo{}
 	if blockData.PoolInfo != nil {
-		tpTarget := exp.ChainParams.TicketPoolSize * exp.ChainParams.TicketsPerBlock
+		tpTarget := uint32(exp.ChainParams.TicketPoolSize) * uint32(exp.ChainParams.TicketsPerBlock)
 		p.HomeInfo.PoolInfo = types.TicketPoolInfo{
 			Size:          blockData.PoolInfo.Size,
 			Value:         blockData.PoolInfo.Value,
